requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
python-multipart==0.0.9
pydantic>=2.11.1
pydantic-settings==2.5.2
Jinja2==3.1.4
openai==1.47.0
google-adk==1.12.0
google-generativeai
python-dotenv==1.0.1
sqlite3